Core: Fixed wchar Filter, FilterWhile
git-svn-id: svn://ultimatepp.org/upp/trunk@1440 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/StrUtil.cpp
@@ -118,7 +118,7 @@ WString Filter(const wchar *s, int (*filter)(int))
 {
 	WString result;
 	while(*s) {
-		int c = (*filter)((byte)*s++);
+		int c = (*filter)(*s++);
 		if(c) result.Cat(c);
 	}
 	return result;
@@ -128,7 +128,7 @@ WString FilterWhile(const wchar *s, int (*filter)(int))
 {
 	WString result;
 	while(*s) {
-		int c = (*filter)((byte)*s++);
+		int c = (*filter)(*s++);
 		if(!c) break;
 		result.Cat(c);
 	}
